agent/requirements.txt
livekit-agents>=1.0.0
livekit-plugins-google>=1.0.0
livekit-plugins-silero>=1.0.0
livekit
python-dotenv
